+++ src/stages/proposal-queued.ts
@@ -23,9 +23,8 @@ import {
   findCallScheduledByTxHash,
   getL2TimelockForGovernor,
 } from "../discovery/timelock-discovery";
-import { serializeCallScheduledDataArray } from "../stages/base";
-import { getBlockTimestamp, failPrepare } from "./base";
-import { StageBuilder } from "./stage-builder";
+import { serializeCallScheduledDataArray, getBlockTimestamp, failPrepare } from "./utils";
+import { StageBuilder } from "./builder";
 
 /**
  * Track proposal queued stage